feat: move character sheet to shared space
