+++ services/character-sheet/src/models/discipline-embeddable.schema.ts
