+++ services/character-sheet/src/models/equipment-embeddable.schema.ts
@@ -1,6 +1,6 @@
 import { Prop, Schema } from '@nestjs/mongoose';
 import { IsEnum, IsString } from '@cats-cradle/validation-schemas';
-import { Slot } from '../../data/gear/slot';
+import { Slot } from '../data/gear/slot';
 
 @Schema()
 export class EquipmentEmbeddable {
