+++ services/character-sheet/src/models/stats-embeddable.schema.ts
